# models/index.js
const Author = require('./author');
const Genre = require('./genre');
const Publisher = require('./publisher');
const Book = require('./book');
const BookAuthors = require('./BookAuthors');
const BookPublishers = require('./BookPublishers');

module.exports = {
  Author,
  Genre,
  Publisher,
  Book,
  BookAuthors,
  BookPublishers,
};